backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
motor==3.3.2
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
telethon==1.32.1
qrcode==7.4.2
pillow==10.1.0
aiofiles==23.2.1
websockets==12.0
httpx==0.25.2
schedule==1.2.0
python-crontab==3.0.0
rich==13.7.0
cryptography==41.0.8
socks==0.0.0
PySocks==1.7.1
asyncio-mqtt==0.16.1